Thông báo khi thêm danh mục
chạy migration , thêm pusher key các thứ

+++ database/migrations/2025_02_21_090935_add_role_id_to_users_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->foreignId('role_id')->default(3)->after('email')->constrained('roles')->onDelete('cascade');
+            if (!Schema::hasColumn('users', 'role_id')) {
+                $table->foreignId('role_id')->default(3)->after('email')->constrained('roles')->onDelete('cascade');
+            }
         });
     }
